Changes after code review.
