+++ pubnub-users/src/main/kotlin/com/pubnub/entities/UserExtension.kt
@@ -8,9 +8,9 @@ import com.pubnub.api.enums.PNOperationType
 import com.pubnub.api.models.consumer.objects.PNKey
 import com.pubnub.api.models.consumer.objects.PNPage
 import com.pubnub.api.models.consumer.objects.PNSortKey
-import com.pubnub.api.models.consumer.objects.ResultKey
 import com.pubnub.api.models.consumer.objects.ResultSortKey
 import com.pubnub.entities.models.consumer.user.RemoveUserResult
+import com.pubnub.entities.models.consumer.user.UserKey
 import com.pubnub.entities.models.consumer.user.UserResult
 import com.pubnub.entities.models.consumer.user.UsersResult
 import com.pubnub.entities.models.consumer.user.toRemoveUserResult
@@ -40,7 +40,7 @@ fun PubNub.fetchUsers(
     limit: Int? = null,
     page: PNPage? = null,
     filter: String? = null,
-    sort: Collection<ResultSortKey<ResultKey>> = listOf(),
+    sort: Collection<ResultSortKey<UserKey>> = listOf(),
     includeCount: Boolean = false,
     includeCustom: Boolean = false
 ): ExtendedRemoteAction<UsersResult?> = firstDo(
@@ -59,7 +59,7 @@ fun PubNub.fetchUsers(
     ) { pnUuidMetadataArrayResult -> pnUuidMetadataArrayResult.toUsersResult() }
 }
 
-private fun toPNSortKey(sort: Collection<ResultSortKey<ResultKey>>): Collection<PNSortKey<PNKey>> {
+private fun toPNSortKey(sort: Collection<ResultSortKey<UserKey>>): Collection<PNSortKey<PNKey>> {
     val pnSortKeyList = sort.map { resultSortKey ->
         when (resultSortKey) {
             is ResultSortKey.Asc -> PNSortKey.PNAsc(key = resultSortKey.key.toPNSortKey())
